#SB-285: Rename the strongbox-authentication-rdbms to strongbox-authentication-jdbc
Renamed the module.
