+++ strongbox-authentication/strongbox-authentication-jdbc/src/main/java/org/carlspring/strongbox/dao/jdbc/BaseDBDao.java
@@ -1,0 +1,25 @@
+package org.carlspring.strongbox.dao.jdbc;
+
+import org.carlspring.strongbox.security.jaas.authentication.UserResolutionException;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+/**
+ * @author mtodorov
+ */
+public interface BaseDBDao
+{
+
+    String getTableName();
+
+    Connection getConnection()
+            throws SQLException;
+
+    long count()
+            throws UserResolutionException;
+
+    long count(String whereClause)
+            throws UserResolutionException;
+
+}
