+++ strongbox-authentication/strongbox-authentication-jdbc/src/test/resources/META-INF/properties/jdbc.properties
