+++ strongbox-authentication/strongbox-authentication-jdbc/src/test/resources/etc/conf/security-authentication-jdbc.xml
