+++ strongbox-authentication/strongbox-authentication-jdbc/src/test/resources/login.conf
@@ -1,0 +1,3 @@
+strongbox {
+    org.carlspring.strongbox.security.jaas.jdbc.JDBCLoginModule required;
+};
