+++ strongbox-authentication/strongbox-authentication-jdbc/src/test/resources/sql/derby/01-db.sql
