+++ strongbox-authentication/strongbox-authentication-jdbc/src/test/resources/sql/derby/02-ddl.sql
