--- strongbox-authentication/strongbox-authentication-rdbms/src/test/resources/META-INF/spring/strongbox-authentication-rdbms-context.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:context="http://www.springframework.org/schema/context"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-                           http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
-
-    <context:annotation-config />
-    <context:component-scan base-package="org.carlspring.strongbox" />
-
-    <bean id="usersDaoRdbms" class="org.carlspring.strongbox.dao.rdbms.impl.UsersDaoImpl">
-    </bean>
-
-    <bean id="rolesDaoRdbms" class="org.carlspring.strongbox.dao.rdbms.impl.RolesDaoImpl">
-    </bean>
-
-    <bean name="userResolver" class="org.carlspring.strongbox.security.jaas.rdbms.RDBMSUserRealm">
-        <property name="usersDao" ref="usersDaoRdbms"/>
-    </bean>
-
-</beans>
