--- strongbox-authentication/strongbox-authentication-rdbms/src/test/resources/login.conf
@@ -1,3 +1,0 @@
-strongbox {
-    org.carlspring.strongbox.security.jaas.rdbms.RDBMSLoginModule required;
-};
